tmp commit

+++ img_to_tf_record.py
@@ -101,6 +101,8 @@ class Img2TFRecord(object):
         self.__generate_tf_record_files('train', self.__train_image_with_breed, resize, channel, one_record_max_imgaes)
         self.__generate_tf_record_files('test', self.__test_image_with_breed, resize, channel, one_record_max_imgaes)
 
+        print( "generate_tf_record_files file finish.\n" )
+
 
     def read_train_images_from_tf_records(self, reshape, batch_size):
         records_path = os.path.join(self.__out_dir, 'train_*.tfr')
@@ -151,7 +153,8 @@ class Img2TFRecord(object):
     def __generate_tf_record_files(self, prefix, image_with_breed, resize, channel = 1, one_record_max_imgaes = 1024):
         # delete info file
         info_file = os.path.join(self.__out_dir, prefix + '_record.inf')
-        os.remove(info_file)
+        if os.path.exists(info_file):
+            os.remove(info_file)
 
         # main code
         writer = None
@@ -200,6 +203,8 @@ class Img2TFRecord(object):
 
         writer.close()
         self.__save_record_info_to_file(prefix, current_index - 1, image_amount)
+
+        print( "generate {} record file finish, total {} files.\n".format(prefix, current_index) )
 
 
     def __save_record_info_to_file(self, prefix, record_index, image_amount):
